Divide file to Main and UserActions
